requirements.txt
Django==3.1.3
django-filter==21.1
django-querycount==0.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
environs==9.5.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
#apartir de aqui todo chido
autoflake==2.1.1
black==23.3.0
commitizen==3.2.2
gunicorn==20.1.0
isort==5.12.0
pre-commit==3.3.1